Time: 0 ms (100%), Space: 8.2 MB (99.67%) - LeetHub

+++ 3768-check-if-digits-are-equal-in-string-after-operations-i/3768-check-if-digits-are-equal-in-string-after-operations-i.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     bool hasSameDigits(string& s) {
-        int n = s.size();
+        uint8_t n = s.size();
         for (int i = n - 1; i > 1; i--) {
             for (int j = 0; j < i; j++) {
                 s[j] = (s[j] + s[j + 1]) % 10 + '0';
